WORKING | REFACTOR: LoadingHandler | MetaManager | GameManager
Flow now working through to RESTART button.

Find flow again and then Refactor MetaManager methods to >> MetaUI

We want MetaManager to be handling as little as possible BUT it should be handling the transitions pending conditions met.

Same for GameManager >> handle as little as possible but manage the transitions pending conditions.